Extend Apprenticeship Update to Based class
