+++ src/SFA.DAS.Payments.DataLocks.Domain/Services/Apprenticeships/ApprenticeshipService.cs
@@ -19,10 +19,16 @@ namespace SFA.DAS.Payments.DataLocks.Domain.Services.Apprenticeships
     public class ApprenticeshipService : IApprenticeshipService
     {
         private readonly IApprenticeshipRepository repository;
+        private readonly IApprenticeshipApprovedUpdatedService apprenticeshipApprovedUpdatedService;
+        private readonly IApprenticeshipDataLockTriageService apprenticeshipDataLockTriageService;
 
-        public ApprenticeshipService(IApprenticeshipRepository repository)
+        public ApprenticeshipService(IApprenticeshipRepository repository, 
+            IApprenticeshipApprovedUpdatedService apprenticeshipUpdatedService,
+            IApprenticeshipDataLockTriageService apprenticeshipDataLockTriageService)
         {
             this.repository = repository ?? throw new ArgumentNullException(nameof(repository));
+            this.apprenticeshipApprovedUpdatedService = apprenticeshipUpdatedService;
+            this.apprenticeshipDataLockTriageService = apprenticeshipDataLockTriageService;
         }
 
         public async Task<List<ApprenticeshipDuplicateModel>> NewApprenticeship(ApprenticeshipModel newApprenticeship)
@@ -68,90 +74,16 @@ namespace SFA.DAS.Payments.DataLocks.Domain.Services.Apprenticeships
             }
         }
 
-        public async Task<ApprenticeshipModel> UpdateApprenticeship(UpdatedApprenticeshipApprovedModel updatedApprenticeshipApproved)
+        public async Task<ApprenticeshipModel> UpdateApprenticeship(UpdatedApprenticeshipApprovedModel updatedApprenticeship)
         {
-            var currentApprenticeship = await GetApprenticeship(updatedApprenticeshipApproved.ApprenticeshipId).ConfigureAwait(false);
-
-            using (var scope = new TransactionScope(TransactionScopeOption.Required, TransactionScopeAsyncFlowOption.Enabled))
-            {
-                currentApprenticeship.AgreedOnDate = updatedApprenticeshipApproved.AgreedOnDate;
-                currentApprenticeship.Uln = updatedApprenticeshipApproved.Uln;
-                currentApprenticeship.EstimatedStartDate = updatedApprenticeshipApproved.EstimatedStartDate;
-                currentApprenticeship.EstimatedEndDate = updatedApprenticeshipApproved.EstimatedEndDate;
-                currentApprenticeship.StandardCode = updatedApprenticeshipApproved.StandardCode;
-                currentApprenticeship.ProgrammeType = updatedApprenticeshipApproved.ProgrammeType;
-                currentApprenticeship.FrameworkCode = updatedApprenticeshipApproved.FrameworkCode;
-                currentApprenticeship.PathwayCode = updatedApprenticeshipApproved.PathwayCode;
-
-                UpdatePriceEpisodes(updatedApprenticeshipApproved.ApprenticeshipPriceEpisodes, currentApprenticeship.ApprenticeshipPriceEpisodes);
-
-                await repository.UpdateApprenticeship(currentApprenticeship);
-
-                var latestApprenticeship = await GetApprenticeship(updatedApprenticeshipApproved.ApprenticeshipId).ConfigureAwait(false);
-
-                scope.Complete();
-
-                return latestApprenticeship;
-            }
+            return await apprenticeshipApprovedUpdatedService.UpdateApprenticeship(updatedApprenticeship);
         }
 
         public async Task<ApprenticeshipModel> UpdateApprenticeshipForDataLockTriage(UpdatedApprenticeshipDataLockTriageModel updatedApprenticeship)
         {
-            var currentApprenticeship = await GetApprenticeship(updatedApprenticeship.ApprenticeshipId).ConfigureAwait(false);
-
-            using (var scope = new TransactionScope(TransactionScopeOption.Required, TransactionScopeAsyncFlowOption.Enabled))
-            {
-                currentApprenticeship.AgreedOnDate = updatedApprenticeship.AgreedOnDate;
-                currentApprenticeship.StandardCode = updatedApprenticeship.StandardCode;
-                currentApprenticeship.ProgrammeType = updatedApprenticeship.ProgrammeType;
-                currentApprenticeship.FrameworkCode = updatedApprenticeship.FrameworkCode;
-                currentApprenticeship.PathwayCode = updatedApprenticeship.PathwayCode;
-
-                UpdatePriceEpisodes(updatedApprenticeship.ApprenticeshipPriceEpisodes, currentApprenticeship.ApprenticeshipPriceEpisodes);
-                
-                await repository.UpdateApprenticeship(currentApprenticeship);
-                var latestApprenticeship = await GetApprenticeship(updatedApprenticeship.ApprenticeshipId).ConfigureAwait(false);
-
-                scope.Complete();
-
-                return latestApprenticeship;
-            }
+            return await apprenticeshipDataLockTriageService.UpdateApprenticeship(updatedApprenticeship);
         }
 
-        private async Task<ApprenticeshipModel> GetApprenticeship(long apprenticeshipId)
-        {
-            var currentApprenticeship = await repository.Get(apprenticeshipId);
-            if (currentApprenticeship == null)
-            {
-                throw new InvalidOperationException(
-                    $"Cannot find updated apprenticeship Apprenticeship id: {apprenticeshipId}");
-            }
 
-            return currentApprenticeship;
-        }
-        
-        private void UpdatePriceEpisodes(List<ApprenticeshipPriceEpisodeModel> receivedPriceEpisodes, List<ApprenticeshipPriceEpisodeModel> currentPriceEpisodes)
-        {
-            foreach (var currentPriceEpisode in currentPriceEpisodes)
-            {
-                var updatedPriceEpisode = receivedPriceEpisodes
-                    .SingleOrDefault(o => o.StartDate == currentPriceEpisode.StartDate && o.Cost == currentPriceEpisode.Cost);
-
-                if (updatedPriceEpisode == null)
-                {
-                    currentPriceEpisode.Removed = true;
-                }
-                else
-                {
-                    currentPriceEpisode.EndDate = updatedPriceEpisode.EndDate;
-                }
-            }
-
-            var newPriceEpisodes = receivedPriceEpisodes
-                .Where(x => currentPriceEpisodes.All(o => o.StartDate != x.StartDate && o.Cost != x.Cost))
-                .ToList();
-
-            currentPriceEpisodes.AddRange(newPriceEpisodes);
-        }
     }
 }
