+++ src/SFA.DAS.Payments.DataLocks.Domain/Services/Apprenticeships/IApprenticeshipUpdatedService.cs
@@ -1,0 +1,19 @@
+﻿using System.Threading.Tasks;
+using SFA.DAS.Payments.DataLocks.Domain.Models;
+using SFA.DAS.Payments.Model.Core.Entities;
+
+namespace SFA.DAS.Payments.DataLocks.Domain.Services.Apprenticeships
+{
+    public interface IApprenticeshipUpdatedService<in T> where T : UpdatedApprenticeshipModel
+    {
+        Task<ApprenticeshipModel> UpdateApprenticeship(T updatedApprenticeship);
+    }
+
+    public interface IApprenticeshipApprovedUpdatedService : IApprenticeshipUpdatedService<UpdatedApprenticeshipApprovedModel>
+    {
+    }
+
+    public interface IApprenticeshipDataLockTriageService : IApprenticeshipUpdatedService<UpdatedApprenticeshipDataLockTriageModel>
+    {
+    }
+}
